292: Move io package into services
Consolidates `.eng` file generation under `machwave.services.eng` and removes
the top-level `io` package, which only held that single module and shadowed
the stdlib `io` name. Tests and docs move alongside the code.
