--- docs/api/io.md
@@ -1,5 +1,0 @@
-# io
-
-Import/export utilities for interoperability with external rocket simulation tools. Currently provides `.eng` file generation — the standard thrust-curve format used by OpenRocket, RASAero, and other flight simulators. Supply time-thrust arrays from a Machwave simulation and get a ready-to-use `.eng` string.
-
-::: machwave.io
