+++ docs/api/services/eng.md
@@ -1,0 +1,5 @@
+# eng
+
+Import/export utilities for interoperability with external rocket simulation tools. Provides `.eng` file generation — the standard thrust-curve format used by OpenRocket, RASAero, and other flight simulators. Supply time-thrust arrays from a Machwave simulation and get a ready-to-use `.eng` string.
+
+::: machwave.services.eng
